+++ api/oc_knx_dev.c
@@ -296,14 +296,14 @@ oc_core_dev_ia_put_handler(oc_request_t *request,
       if (rep->iname == 12) {
         PRINT("  oc_core_dev_ia_put_handler received 12 (ia) : %d\n",
               (int)rep->value.integer);
-        oc_core_set_device_ia(device_index, (int)rep->value.integer);
+        oc_core_set_device_ia(device_index, (uint32_t)rep->value.integer);
         int temp = (int)rep->value.integer;
         oc_storage_write(KNX_STORAGE_IA, (uint8_t *)&temp, sizeof(temp));
         ia_set = true;
       } else if (rep->iname == 25) {
         PRINT("  oc_core_dev_ia_put_handler received 25 (fid): %d\n",
               (int)rep->value.integer);
-        oc_core_set_device_fid(device_index, (int)rep->value.integer);
+        oc_core_set_device_fid(device_index, rep->value.integer);
         int temp = (int)rep->value.integer;
         oc_storage_write(KNX_STORAGE_FID, (uint8_t *)&temp, sizeof(temp));
         fid_set = true;
@@ -931,7 +931,7 @@ oc_knx_device_storage_read(size_t device_index)
   /* KNX_STORAGE_IID */
   temp_size =
     oc_storage_read(KNX_STORAGE_IID, (uint8_t *)&device->iid, sizeof(int64_t));
-  PRINT("  idd (storage) %ld\n", device->iid);
+  PRINT("  idd (storage) %lld\n", device->iid);
 
   /* KNX_STORAGE_PM */
   temp_size = oc_storage_read(KNX_STORAGE_PM, (uint8_t *)&pm, 1);
